+++ src/components/filemanagement/ImportJsonFile.js
@@ -5,15 +5,33 @@ import { setContent, clearTabPath } from '../../store/slices/content';
 import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew';
 import { useTranslation } from 'react-i18next';
 import { MuiFileInput } from 'mui-file-input'
+import { addThrow } from '../../store/slices/throws.js';
+import { format } from 'date-fns';
+
 
 const ImportJsonFile = (props) => {
 
   const { t } = useTranslation();
   const dispatch = useDispatch();
 
+  const [jsonSaveFile, setJsonSaveFile] = React.useState(null)
   const [jsonFile, setJsonFile] = React.useState(null)
 
-  const handleChange = async (newValue) => {
+  const handleSingleContentImport = async (newValue) => {
+    setJsonSaveFile(newValue);
+    const jsonText = await newValue.text();
+    const singleSavedContent = JSON.parse(jsonText);
+
+    if (!singleSavedContent.contentData || !singleSavedContent.contentType) {
+      props.returnToMenu();
+      return;
+    }
+
+    dispatch(addThrow({ result: singleSavedContent.contentData, timestamp: format(new Date(), "yyyy-MM-dd' 'HH:mm:ss") }));
+    props.returnToMenu();
+  }
+
+  const handleSandboxImport = async (newValue) => {
     setJsonFile(newValue);
     const jsonText = await newValue.text();
     dispatch(clearTabPath());
@@ -32,14 +50,17 @@ const ImportJsonFile = (props) => {
       </Grid>
       <Grid item xs={12}>&nbsp;</Grid>
       <Grid item xs={12}>
-        <Typography>{t("Tap to select a JSON file to import")}</Typography>
+        <Typography>Importazione di un singolo contenuto (schede personaggio, mappe, dungeon..). {t("Tap to select a JSON file to import")}.</Typography>
       </Grid>
       <Grid item xs={12}>
-        <MuiFileInput style={{width: '100%'}} getInputText={(value) => value ? value.name : t('Tap to select a file to import') } inputProps={{ accept: '.json' }} value={jsonFile} onChange={handleChange} />
+        <MuiFileInput style={{width: '100%'}} getInputText={(value) => value ? value.name : t('Tap to select a file to import') } inputProps={{ accept: '.json' }} value={jsonSaveFile} onChange={handleSingleContentImport} />
       </Grid>
       <Grid item xs={12}>&nbsp;</Grid>
       <Grid item xs={12}>
-        <Typography>{t("Warning: ALL existing content will be overwritten!")}</Typography>
+        <Typography>Importazione di una intera sandbox. {t("Tap to select a JSON file to import")}. {t("Warning: ALL existing content will be overwritten!")}</Typography>
+      </Grid>
+      <Grid item xs={12}>
+        <MuiFileInput style={{width: '100%'}} getInputText={(value) => value ? value.name : t('Tap to select a file to import') } inputProps={{ accept: '.json' }} value={jsonFile} onChange={handleSandboxImport} />
       </Grid>
       <Grid item xs={12}>&nbsp;</Grid>
     </Grid>