+++ Changelog.md
@@ -1,8 +1,12 @@
 # Changelog Template progetto DM Whisper
 
+    18 Dic 2024
+    -----------
+    - 1.54.0 You can now save and load saved content, of any kind.
+
     17 Dic 2024
     -----------
-    - 1.57.0 Export saved content to PDF.
+    - 1.53.0 Export saved content to PDF.
 
     16 Dic 2024
     -----------